@@ -1,17 +1,16 @@
 module github.com/Azure/aks-secure-tls-bootstrap/service
 
-go 1.21
+go 1.23.6
 
 require (
-	go.uber.org/mock v0.2.0
-	google.golang.org/grpc v1.60.1
-	google.golang.org/protobuf v1.33.0
+	go.uber.org/mock v0.5.0
+	google.golang.org/grpc v1.70.0
+	google.golang.org/protobuf v1.35.2
 )
 
 require (
-	github.com/golang/protobuf v1.5.3 // indirect
 	golang.org/x/net v0.33.0 // indirect
 	golang.org/x/sys v0.28.0 // indirect
 	golang.org/x/text v0.21.0 // indirect
-	google.golang.org/genproto/googleapis/rpc v0.0.0-20231002182017-d307bd883b97 // indirect
+	google.golang.org/genproto/googleapis/rpc v0.0.0-20241202173237-19429a94021a // indirect
 )
