@@ -2,7 +2,12 @@
 // Licensed under the MIT license.
 
 // Code generated by MockGen. DO NOT EDIT.
-// Source: service/protos/bootstrap_grpc.pb.go
+// Source: pkg/gen/akssecuretlsbootstrap/v1/service_grpc.pb.go
+//
+// Generated by this command:
+//
+//	mockgen -package=mocks -copyright_file=/Users/cam/repos/aks-secure-tls-bootstrap/service/../hack/copyright_header.txt -source=pkg/gen/akssecuretlsbootstrap/v1/service_grpc.pb.go -destination=pkg/mocks/akssecuretlsbootstrap/v1/service.go github.com/Azure/aks-secure-tls-bootstrap/service/proto SecureTLSBootstrapServiceClient
+//
 
 // Package mocks is a generated GoMock package.
 package mocks
@@ -11,7 +16,7 @@ import (
 	context "context"
 	reflect "reflect"
 
-	protos "github.com/Azure/aks-secure-tls-bootstrap/service/protos"
+	akssecuretlsbootstrapv1 "github.com/Azure/aks-secure-tls-bootstrap/service/pkg/gen/akssecuretlsbootstrap/v1"
 	gomock "go.uber.org/mock/gomock"
 	grpc "google.golang.org/grpc"
 )
@@ -20,6 +25,7 @@ import (
 type MockSecureTLSBootstrapServiceClient struct {
 	ctrl     *gomock.Controller
 	recorder *MockSecureTLSBootstrapServiceClientMockRecorder
+	isgomock struct{}
 }
 
 // MockSecureTLSBootstrapServiceClientMockRecorder is the mock recorder for MockSecureTLSBootstrapServiceClient.
@@ -40,69 +46,50 @@ func (m *MockSecureTLSBootstrapServiceClient) EXPECT() *MockSecureTLSBootstrapSe
 }
 
 // GetCredential mocks base method.
-func (m *MockSecureTLSBootstrapServiceClient) GetCredential(ctx context.Context, in *protos.CredentialRequest, opts ...grpc.CallOption) (*protos.CredentialResponse, error) {
+func (m *MockSecureTLSBootstrapServiceClient) GetCredential(ctx context.Context, in *akssecuretlsbootstrapv1.GetCredentialRequest, opts ...grpc.CallOption) (*akssecuretlsbootstrapv1.GetCredentialResponse, error) {
 	m.ctrl.T.Helper()
-	varargs := []interface{}{ctx, in}
+	varargs := []any{ctx, in}
 	for _, a := range opts {
 		varargs = append(varargs, a)
 	}
 	ret := m.ctrl.Call(m, "GetCredential", varargs...)
-	ret0, _ := ret[0].(*protos.CredentialResponse)
+	ret0, _ := ret[0].(*akssecuretlsbootstrapv1.GetCredentialResponse)
 	ret1, _ := ret[1].(error)
 	return ret0, ret1
 }
 
 // GetCredential indicates an expected call of GetCredential.
-func (mr *MockSecureTLSBootstrapServiceClientMockRecorder) GetCredential(ctx, in interface{}, opts ...interface{}) *gomock.Call {
+func (mr *MockSecureTLSBootstrapServiceClientMockRecorder) GetCredential(ctx, in any, opts ...any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	varargs := append([]interface{}{ctx, in}, opts...)
+	varargs := append([]any{ctx, in}, opts...)
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "GetCredential", reflect.TypeOf((*MockSecureTLSBootstrapServiceClient)(nil).GetCredential), varargs...)
 }
 
 // GetNonce mocks base method.
-func (m *MockSecureTLSBootstrapServiceClient) GetNonce(ctx context.Context, in *protos.NonceRequest, opts ...grpc.CallOption) (*protos.NonceResponse, error) {
+func (m *MockSecureTLSBootstrapServiceClient) GetNonce(ctx context.Context, in *akssecuretlsbootstrapv1.GetNonceRequest, opts ...grpc.CallOption) (*akssecuretlsbootstrapv1.GetNonceResponse, error) {
 	m.ctrl.T.Helper()
-	varargs := []interface{}{ctx, in}
+	varargs := []any{ctx, in}
 	for _, a := range opts {
 		varargs = append(varargs, a)
 	}
 	ret := m.ctrl.Call(m, "GetNonce", varargs...)
-	ret0, _ := ret[0].(*protos.NonceResponse)
+	ret0, _ := ret[0].(*akssecuretlsbootstrapv1.GetNonceResponse)
 	ret1, _ := ret[1].(error)
 	return ret0, ret1
 }
 
 // GetNonce indicates an expected call of GetNonce.
-func (mr *MockSecureTLSBootstrapServiceClientMockRecorder) GetNonce(ctx, in interface{}, opts ...interface{}) *gomock.Call {
+func (mr *MockSecureTLSBootstrapServiceClientMockRecorder) GetNonce(ctx, in any, opts ...any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	varargs := append([]interface{}{ctx, in}, opts...)
+	varargs := append([]any{ctx, in}, opts...)
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "GetNonce", reflect.TypeOf((*MockSecureTLSBootstrapServiceClient)(nil).GetNonce), varargs...)
-}
-
-// GetToken mocks base method.
-func (m *MockSecureTLSBootstrapServiceClient) GetToken(ctx context.Context, in *protos.TokenRequest, opts ...grpc.CallOption) (*protos.TokenResponse, error) {
-	m.ctrl.T.Helper()
-	varargs := []interface{}{ctx, in}
-	for _, a := range opts {
-		varargs = append(varargs, a)
-	}
-	ret := m.ctrl.Call(m, "GetToken", varargs...)
-	ret0, _ := ret[0].(*protos.TokenResponse)
-	ret1, _ := ret[1].(error)
-	return ret0, ret1
-}
-
-// GetToken indicates an expected call of GetToken.
-func (mr *MockSecureTLSBootstrapServiceClientMockRecorder) GetToken(ctx, in interface{}, opts ...interface{}) *gomock.Call {
-	mr.mock.ctrl.T.Helper()
-	varargs := append([]interface{}{ctx, in}, opts...)
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "GetToken", reflect.TypeOf((*MockSecureTLSBootstrapServiceClient)(nil).GetToken), varargs...)
 }
 
 // MockSecureTLSBootstrapServiceServer is a mock of SecureTLSBootstrapServiceServer interface.
 type MockSecureTLSBootstrapServiceServer struct {
 	ctrl     *gomock.Controller
 	recorder *MockSecureTLSBootstrapServiceServerMockRecorder
+	isgomock struct{}
 }
 
 // MockSecureTLSBootstrapServiceServerMockRecorder is the mock recorder for MockSecureTLSBootstrapServiceServer.
@@ -123,48 +110,33 @@ func (m *MockSecureTLSBootstrapServiceServer) EXPECT() *MockSecureTLSBootstrapSe
 }
 
 // GetCredential mocks base method.
-func (m *MockSecureTLSBootstrapServiceServer) GetCredential(arg0 context.Context, arg1 *protos.CredentialRequest) (*protos.CredentialResponse, error) {
+func (m *MockSecureTLSBootstrapServiceServer) GetCredential(arg0 context.Context, arg1 *akssecuretlsbootstrapv1.GetCredentialRequest) (*akssecuretlsbootstrapv1.GetCredentialResponse, error) {
 	m.ctrl.T.Helper()
 	ret := m.ctrl.Call(m, "GetCredential", arg0, arg1)
-	ret0, _ := ret[0].(*protos.CredentialResponse)
+	ret0, _ := ret[0].(*akssecuretlsbootstrapv1.GetCredentialResponse)
 	ret1, _ := ret[1].(error)
 	return ret0, ret1
 }
 
 // GetCredential indicates an expected call of GetCredential.
-func (mr *MockSecureTLSBootstrapServiceServerMockRecorder) GetCredential(arg0, arg1 interface{}) *gomock.Call {
+func (mr *MockSecureTLSBootstrapServiceServerMockRecorder) GetCredential(arg0, arg1 any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "GetCredential", reflect.TypeOf((*MockSecureTLSBootstrapServiceServer)(nil).GetCredential), arg0, arg1)
 }
 
 // GetNonce mocks base method.
-func (m *MockSecureTLSBootstrapServiceServer) GetNonce(arg0 context.Context, arg1 *protos.NonceRequest) (*protos.NonceResponse, error) {
+func (m *MockSecureTLSBootstrapServiceServer) GetNonce(arg0 context.Context, arg1 *akssecuretlsbootstrapv1.GetNonceRequest) (*akssecuretlsbootstrapv1.GetNonceResponse, error) {
 	m.ctrl.T.Helper()
 	ret := m.ctrl.Call(m, "GetNonce", arg0, arg1)
-	ret0, _ := ret[0].(*protos.NonceResponse)
+	ret0, _ := ret[0].(*akssecuretlsbootstrapv1.GetNonceResponse)
 	ret1, _ := ret[1].(error)
 	return ret0, ret1
 }
 
 // GetNonce indicates an expected call of GetNonce.
-func (mr *MockSecureTLSBootstrapServiceServerMockRecorder) GetNonce(arg0, arg1 interface{}) *gomock.Call {
+func (mr *MockSecureTLSBootstrapServiceServerMockRecorder) GetNonce(arg0, arg1 any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "GetNonce", reflect.TypeOf((*MockSecureTLSBootstrapServiceServer)(nil).GetNonce), arg0, arg1)
-}
-
-// GetToken mocks base method.
-func (m *MockSecureTLSBootstrapServiceServer) GetToken(arg0 context.Context, arg1 *protos.TokenRequest) (*protos.TokenResponse, error) {
-	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "GetToken", arg0, arg1)
-	ret0, _ := ret[0].(*protos.TokenResponse)
-	ret1, _ := ret[1].(error)
-	return ret0, ret1
-}
-
-// GetToken indicates an expected call of GetToken.
-func (mr *MockSecureTLSBootstrapServiceServerMockRecorder) GetToken(arg0, arg1 interface{}) *gomock.Call {
-	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "GetToken", reflect.TypeOf((*MockSecureTLSBootstrapServiceServer)(nil).GetToken), arg0, arg1)
 }
 
 // mustEmbedUnimplementedSecureTLSBootstrapServiceServer mocks base method.
@@ -183,6 +155,7 @@ func (mr *MockSecureTLSBootstrapServiceServerMockRecorder) mustEmbedUnimplemente
 type MockUnsafeSecureTLSBootstrapServiceServer struct {
 	ctrl     *gomock.Controller
 	recorder *MockUnsafeSecureTLSBootstrapServiceServerMockRecorder
+	isgomock struct{}
 }
 
 // MockUnsafeSecureTLSBootstrapServiceServerMockRecorder is the mock recorder for MockUnsafeSecureTLSBootstrapServiceServer.
